Updated name/version, Misc restructuring, publish to npm
