+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "",
-  "version": "0.0.0",
+  "name": "@pearson-incubator/bookshelf",
+  "version": "0.1.0",
   "description": "",
   "main": "./main.js",
   "scripts": {
@@ -14,8 +14,7 @@
     "version": "npm run gen-changelog && git add CHANGELOG.md",
     "ci": "npm test && npm run build",
     "release": "node ./npm_scripts/release.js",
-    "verify": "node ./npm_scripts/verify.js",
-    "postinstall": "cp ./node_modules/@pearson-components/npm-scripts/npm_scripts/*.js ./npm_scripts"
+    "verify": "node ./npm_scripts/verify.js"
   },
   "author": "",
   "license": "MIT",
