+++ webpack.config.js
@@ -47,7 +47,7 @@ module.exports = {
         loader: 'style!css!sass' // sass -> css -> javascript -> inline style
       },
       {
-        test: /\.js$/,
+        test: /\.(js|jsx)$/,
         loader: 'babel',
         query: {
           cacheDirectory: true,
@@ -63,5 +63,8 @@ module.exports = {
         loader: 'url-loader?limit=10000' 
       }
     ]
+  },
+  resolve: {
+    extensions: ['', '.webpack.js', '.web.js', '.js', '.jsx']
   }
 };
